Needs further work in the furute
This was code to try to help identify which button had been pressed so that synthesised speech could be played on pressing the button.
This didn't quite work as it would only work after the button had been pressed i.e. the synthesised speech would not play first time.
A workaround was found involving looking at deeper properties of the button pressed.