Update TotemController.cpp
New code to rescale the pole and model when objects are moved.
This avoids the 'long pole' isuue which existed previously, whereby adding objects to the pole and then moving them down led to a long pole and a distant view.  The pole now shortens and lengthens dynamically which provides a closer view of the model at all times.